+++ src/procfunc/random.py
@@ -75,6 +75,7 @@ def clip_gaussian(
     )
 
 
+@pf.tracer.primitive(allow_exec=True)
 def wrap_gaussian(
     rng: np.random.Generator,
     mean: float,
@@ -92,6 +93,7 @@ def wrap_gaussian(
     return x
 
 
+@pf.tracer.primitive(allow_exec=True)
 def exponential(
     rng: np.random.Generator,
     scale: float,
@@ -99,6 +101,7 @@ def exponential(
     return rng.exponential(scale)
 
 
+@pf.tracer.primitive(allow_exec=True)
 def log_uniform(
     rng: np.random.Generator,
     low: float,
@@ -108,6 +111,7 @@ def log_uniform(
     return np.exp(rng.uniform(np.log(low), np.log(high), size=size))
 
 
+@pf.tracer.primitive(allow_exec=True)
 def log_normal(
     rng: np.random.Generator,
     mean: float,
@@ -117,6 +121,7 @@ def log_normal(
     return np.exp(rng.normal(np.log(mean), std, size=size))
 
 
+@pf.tracer.primitive(allow_exec=True)
 def spherical_sample(
     rng: np.random.Generator,
     min_elevation: float | None = None,
@@ -137,6 +142,7 @@ def spherical_sample(
     )
 
 
+@pf.tracer.primitive(allow_exec=True)
 def mixture_of_gaussian(
     rng: np.random.Generator,
     means: np.ndarray,
@@ -161,6 +167,7 @@ def mixture_of_gaussian(
     return res
 
 
+@pf.tracer.primitive(allow_exec=True)
 def beta(
     rng: np.random.Generator,
     a: float,
@@ -169,6 +176,7 @@ def beta(
     return rng.beta(a, b)
 
 
+@pf.tracer.primitive(allow_exec=True)
 def poisson(
     rng: np.random.Generator,
     lam: float,
@@ -176,6 +184,7 @@ def poisson(
     return rng.poisson(lam)
 
 
+@pf.tracer.primitive(allow_exec=True)
 def triangular(
     rng: np.random.Generator,
     low: float,
@@ -185,6 +194,7 @@ def triangular(
     return rng.triangular(low, mode, high)
 
 
+@pf.tracer.primitive(allow_exec=True)
 def gamma(
     rng: np.random.Generator,
     shape: float,
@@ -193,6 +203,7 @@ def gamma(
     return rng.gamma(shape, scale)
 
 
+@pf.tracer.primitive(allow_exec=True)
 def binomial(
     rng: np.random.Generator,
     n: int,
@@ -201,6 +212,7 @@ def binomial(
     return rng.binomial(n, p)
 
 
+@pf.tracer.primitive(allow_exec=True)
 def geometric(
     rng: np.random.Generator,
     p: float,